requirements-dev.txt
# Development dependencies
ruff==0.4.1
pytest==9.0.1
pytest-cov==6.0.0
pytest-django==4.11.1
coverage==7.6.1
pre-commit==3.6.0
django-debug-toolbar==4.4.6
